Code Product

+++ BE/DreamyShop.Api/Controllers/ProductController.cs
@@ -1,4 +1,5 @@
 ﻿using DreamyShop.Domain.Shared.Dtos;
+using DreamyShop.Logic.Conditions;
 using DreamyShop.Logic.Product;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -31,6 +32,30 @@ namespace DreamyShop.Api.Controllers
                 return BadRequest(ModelState);
             }
             var result = await _productService.CreateProduct(productCreateUpdateDto);
+            return Ok(result);
+        }
+
+        [HttpPut("product/updateProduct")]
+        [Authorize]
+        public async Task<IActionResult> UpdateProduct(Guid id, [FromForm] ProductCreateUpdateDto productCreateUpdateDto)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var result = await _productService.UpdateProduct(id, productCreateUpdateDto);
+            return Ok(result);
+        }
+
+        [HttpDelete("product/removeProduct")]
+        [Authorize]
+        public async Task<IActionResult> RemoveAtributeProduct(Guid id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var result = await _productService.RemoveProduct(id);
             return Ok(result);
         }
 
@@ -70,7 +95,26 @@ namespace DreamyShop.Api.Controllers
         [Authorize]
         public async Task<IActionResult> RemoveAtributeProduct(Guid attributeId, Guid attributeTypeId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var result = await _productService.RemoveAtributeProduct(attributeId, attributeTypeId);
+            return Ok(result);
+        }
+
+
+        [HttpPut("product/searchCondition")]
+        [Authorize]
+        public async Task<IActionResult> SearchProduct([FromForm] SearchProductCondition searchProductCondition)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var result = await _productService.SearchProduct(searchProductCondition);
+            if (result.Result == null)
+            {
+                return BadRequest(result);
+            }
             return Ok(result);
         }
     }
